Initial commit of original spriteditor.com source code
Original commit of the index.html file for the Sam Coupe' sprite editor formerly reachable at www.spriteeditor.com

The software is over 10 years old and is presented 'as was' rather than 'as is'. I don't intend on changing it in this moment or the next so have the heart to accept the atrocious use of java script for what it was; a means to allow those with better intentions to continue developing software in support of retro computing platforms they continue to love and respect.

To the best of my knowledge this software remains the only 8 bit sprite editor available online or otherwise with angular rotation and % scaling functionality.

This code is dedicated to my friend Chris 'You don't need to create a new sprite editor!' Pile.